+++ crates/primitives/src/lib.rs
@@ -7,19 +7,19 @@
 
 //! Commonly used types in reth.
 
+mod account;
 mod block;
 mod header;
 mod log;
 mod receipt;
 mod transaction;
-mod account;
 
+pub use account::Account;
 pub use block::{Block, BlockLocked};
 pub use header::{Header, HeaderLocked};
 pub use log::Log;
 pub use receipt::Receipt;
 pub use transaction::{AccessList, AccessListItem, Transaction, TransactionSigned, TxType};
-pub use account::Account;
 
 /// Block Number is height of chain
 pub type BlockNumber = u64;
